backend/engines/audio_engine.py
import librosa
import soundfile as sf
import numpy as np
from pathlib import Path
import subprocess
import sys
import imagehash
from PIL import Image
import os
import time
import subprocess
import tempfile
from typing import Dict, List, Tuple

class AudioHashEngine:
    """
    Sentinel Audio Fingerprinting Engine.
    Generates perceptual hashes from audio spectrograms for piracy detection.
    Uses mel-spectrogram analysis for robust audio matching.
    """
    
    def __init__(self, sample_rate: int = 22050, n_mels: int = 128, 
                 hop_length: int = 512, hash_size: int = 8, chunk_duration: int = 5):
        """
        Initialize the audio hash engine.
        
        Args:
            sample_rate: Audio sampling rate (default: 22050 Hz)
            n_mels: Number of mel bands (default: 128)
            hop_length: Hop length for STFT (default: 512)
            hash_size: Size of the pHash matrix (default: 8 = 64-bit hash)
            chunk_duration: Duration of each audio chunk in seconds (default: 5)
        """
        self.sample_rate = sample_rate
        self.n_mels = n_mels
        self.hop_length = hop_length
        self.hash_size = hash_size
        self.chunk_duration = chunk_duration

    def extract_audio_hash(self, video_path: str, duration: float = None) -> str:
        """
        Extract audio from video, generate a spectrogram pHash.
        
        Args:
            video_path: Path to the video file
            duration: Max duration to process in seconds
            
        Returns:
            Hex string representation of the audio pHash
        """
        if not os.path.exists(video_path):
            raise FileNotFoundError(f"Video file not found: {video_path}")

        try:
            # Load audio using librosa
            y, sr = librosa.load(video_path, sr=self.sample_rate, duration=duration)
            
            # Generate Mel-spectrogram
            S = librosa.feature.melspectrogram(y=y, sr=sr, n_mels=self.n_mels, 
                                             hop_length=self.hop_length)
            
            # Convert to log scale (dB)
            S_db = librosa.power_to_db(S, ref=np.max)
            
            # Normalize to 0-255 for image conversion
            S_norm = ((S_db - S_db.min()) / (S_db.max() - S_db.min()) * 255).astype(np.uint8)
            
            # Convert spectrogram array to PIL Image
            img = Image.fromarray(S_norm)
            
            # Generate pHash from the spectrogram image
            audio_hash = imagehash.phash(img, hash_size=self.hash_size)
            
            return str(audio_hash)
            
        except Exception as e:
            print(f"⚠️ Audio extraction failed (ffmpeg not available): {e}")
            print("   Using fallback hash for demo purposes")
            # Return a deterministic hash based on file path for demo
            import hashlib
            return hashlib.md5(str(video_path).encode()).hexdigest()[:16]

    def _generate_audio_hash(self, audio_data: np.ndarray, sample_rate: int) -> str:
        """Generate perceptual hash from audio spectrogram."""
        try:
            # Create mel-spectrogram
            spec = librosa.feature.melspectrogram(
                y=audio_data, 
                sr=sample_rate, 
                n_mels=self.n_mels
            )
            
            # Convert to decibels
            spec_db = librosa.power_to_db(spec, ref=np.max)
            
            # Normalize to 0-255 range
            spec_norm = ((spec_db - spec_db.min()) / (spec_db.max() - spec_db.min()) * 255).astype(np.uint8)
            
            # Convert to PIL Image
            img = Image.fromarray(spec_norm)
            
            # Generate pHash from the spectrogram image
            audio_hash = imagehash.phash(img, hash_size=self.hash_size)
            
            return str(audio_hash)
            
        except Exception as e:
            print(f"❌ Audio hash generation failed: {e}")
            return None

    def _extract_audio_temp(self, video_path: str) -> str:
        """Extract audio to temporary WAV file using ffmpeg."""
        try:
            # Create temp file
            with tempfile.NamedTemporaryFile(suffix='.wav', delete=False) as temp_file:
                temp_path = temp_file.name
            
            # Find ffmpeg
            ffmpeg_path = r"C:\Users\rishi\AppData\Local\Microsoft\WinGet\Links\ffmpeg.exe"
            if not os.path.exists(ffmpeg_path):
                ffmpeg_path = "ffmpeg"
            
            # Extract audio
            cmd = [
                ffmpeg_path, '-i', video_path,
                '-vn', '-acodec', 'pcm_s16le', 
                '-ar', str(self.sample_rate),
                '-y', temp_path
            ]
            subprocess.run(cmd, check=True, capture_output=True)
            
            return temp_path
            
        except Exception as e:
            print(f"Audio extraction failed: {e}")
            return None

    def hash_audio(self, video_path: str) -> Tuple[List[str], Dict]:
        """
        Generate audio fingerprints for entire video.
        Returns list of hashes and metadata.
        
        Args:
            video_path: Path to video file
            
        Returns:
            (hashes, metadata) tuple
        """
        if not os.path.exists(video_path):
            raise FileNotFoundError(f"Video not found: {video_path}")
        
        start_time = time.time()
        hashes = []
        
        try:
            # Extract audio to temp file first
            temp_audio_path = self._extract_audio_temp(video_path)
            if not temp_audio_path:
                return [], {'duration_seconds': 0, 'processing_time_seconds': 0, 'hash_count': 0}
            
            # Get total duration from extracted audio
            total_duration = librosa.get_duration(path=temp_audio_path)
            
            # Process in chunks
            for start in range(0, int(total_duration), self.chunk_duration):
                try:
                    # Load audio chunk from extracted file
                    y, sr = librosa.load(
                        temp_audio_path, 
                        sr=self.sample_rate,
                        offset=start, 
                        duration=self.chunk_duration
                    )
                    
                    if len(y) == 0:
                        continue
                    
                    # Generate hash for this chunk
                    chunk_hash = self._generate_audio_hash(y, sr)
                    if chunk_hash:
                        hashes.append({
                            'hash': chunk_hash,
                            'timestamp': start
                        })
                        
                except Exception as e:
                    print(f"⚠ Chunk processing failed at {start}s: {e}")
                    continue
            
            # Clean up temp file
            try:
                os.remove(temp_audio_path)
            except:
                pass
            
        except Exception as e:
            print(f"❌ Audio processing failed: {e}")
            return [], {'duration_seconds': 0, 'processing_time_seconds': 0, 'hash_count': 0}
        
        processing_time = time.time() - start_time
        
        metadata = {
            'duration_seconds': total_duration if 'total_duration' in locals() else 0,
            'processing_time_seconds': processing_time,
            'hash_count': len(hashes),
            'chunk_duration': self.chunk_duration,
            'sample_rate': self.sample_rate
        }
        
        return hashes, metadata

if __name__ == "__main__":
    # Quick Test
    engine = AudioHashEngine()
    test_video = "assets/videos/Race Highlights  2026 Australian Grand Prix - FORMULA 1 (720p, h264, youtube).mp4"
    if os.path.exists(test_video):
        print(f"🔍 Testing Audio Hash Engine on: {os.path.basename(test_video)}")
        h = engine.extract_audio_hash(test_video)
        print(f"✅ Audio Fingerprint: {h}")
    else:
        print("❌ Test video not found.")
