backend/requirements.txt
flask
flask-sock
opencv-python
imagehash
pillow
numpy
reportlab
groq
librosa
redis
python-dotenv
flask-cors
requests
soundfile
fakeredis
yt-dlp